Change the meaning of OFFSET parameter in slice.sv
Change the meaning of the OFFSET parameter such that (OFFSET == 0) now only
needs one pixel from the first cycle, (OFFSET == 1) is has two pixels from
the first clock, (OFFSET == 2) has 3 pixels and so on until the largest OFFSET
value possible has all the pixels it needs to do the calculation sourced from
the 1st clock. The mental model to use is 'the larger the OFFSET the larger the
number of pixels are used from the first clock'.

To simplify the use of the slice module, the image bus reordering has been
moved out of the slice module and moved into the engine.